Refactor IntegrationTest structure
  TLS1
